// src/app/core/models/user.ts
export interface User {
  id: string;
  first_name: string
  last_name: string
  email: string
  phone: string
  username: string
  middle_name: any
  preferred_name: string
  profile_picture: string
  date_of_birth: any
  gender: any
  marital_status: any
  nationality: any
  language: any
  occupation: any
  status: string
  type: string
  join_date: any
  baptism_date: any
  confirmation_date: any
  membership_class_date: any
  previous_church: any
  phone_home: any
  phone_mobile: any
  phone_work: any
  address_street: any
  address_city: any
  address_state: any
  address_postal: any
  address_country: any
  emergency_contact_name: any
  emergency_contact_relationship: any
  emergency_contact_phone: any
  emergency_contact_email: any
  ministry_roles: any
  spiritual_gifts: any
  skills: any
  interests: any
  volunteer_preferences: any
  preferred_contact_method: any
  allow_phone_calls: boolean
  allow_emails: boolean
  allow_texts: boolean
  allow_mail: boolean
  is_private: boolean
  share_contact_info: boolean
  share_photos: boolean
  directory_listed: boolean
  family_id: any
  family_role: any
  family_notes: any
  photo_url: any
  bio: any
  notes: any
  special_needs: any
  dietary_restrictions: any
  allergies: any
  custom_fields: any
  tags: any
  tenant_id: string
  member_id: string
  last_attendance_date: any
  attendance_streak: number
  total_attendances: number
  engagement_score: any
  last_contribution_date: any
  last_activity_date: any
  inactive_reason: any
  reactivation_date: any
  last_updated_by_id: any
  last_status_change: any
  status_change_reason: any
  created_at: string
  updated_at: any
}